Write updated changelog to `buildSetup.outputDir`
Because the "Release to GitHub" task handles changelog commits,
we should never commit changelog changes directly, so we need to
revert `CHANGELOG.TXT` in the working tree each time it's updated.
By copying the original to `outputDir` on every build and writing
there instead, we no longer need to worry about reverting OR
duplicate entries.