Fixes incorrect reporting of the language in Windows
Changed using wx.local to ctypes to get the language.

Example of Bug:
Say you live in germany but are English speaking. So you have the system locale set to Germany but the UI displayed language is in English. Using wx.locale would report back as de_DE even if the UI is set to en_US.

This fix reports the current displayed Windows GUI language. Not the system locale language.