Updates PluginInstall.Py
Adds more informative message when running the plugin installer. Messages for Upgrading plugin, Downgrading plugin, Overwriting plugin, Plugin exists but the name is different as well as Install plugin

Adds a GUID check when exporting, if one does not exist a dialog will ask to create one. if one is created it will modify the __init__.py and add a newly generated one. It WILL NOT allow a plugin to be exported without a GUID.

Adds a backup routine for the old version of the plugin if one is installed. it will append the old version to the file name when backing up

Adds a file name suggestion when exporting that has the version of the plugin included in the file name

Adds exception handling for removing the old plugin directory if the plugin directory resides in the eg.corepluginDir it will post a message in the log letting the user know they have to go and manually remove it. otherwise raise the exception