Enables deleting of config data
When removing a plugin eg.config is scanned for any eg.PersistantData the plugin may have set. and if it finds any prompts the user if they would like to delete this data. If yes the data is marked for deletion. If the user does not save their EventGhost session all of the marks for deletion are removed. it also handles the ImportAll() when the debug level is set from a command line.