Fixes bugs for tray icon and empty log
Fixes the taskbar icon removal when show tray icon is not selected.

Fixes the empty log when maxamizing the frame

EG was set to destroy the frame when minimized. this would cause both of the bugs. because the frame would be destryoed so would the logCtrl, therefore deleting any data that may have been in the log. It would also remove the icon from the taskbar because the frame no longer exists.

I Have changed it to hide the frame if the tray icon is to be displayed and to let windows do it's windows thing if not. This doesn't destroy the logCtrl keeping the log data in tact