Fix issue with strings in Speech plugin TextToSpeech action
A change in EventGhost starting with 0.5.0-beta1 seems to have changed how the
string class is handled, breaking the Speech plugin.

Previous to this change attempting to add the TextToSpeech action fails
with:
"AttributeError: class TextToSpeech has no attribute 'buttonInsertTime'"

Previous to this change the TextToSpeech action did not display the correct
name and had no description.

The reason I've done the baseText = Text instead of creating a TextToSpeech
string class as recommended in the internationalisation documentation is the
Speaker and TextToSpeech classes both use the errorCreate string and this
allows them to share the same string instead of having two duplicate strings
to maintain.

This change appears to make
https://github.com/EventGhost/EventGhost/commit/5f76f85d8978786fc7068123924e50ba164c32d6
unnecessary in that the infinite loop no longer occurs event with Text instead
of Text().