[Speech plugin] Fix issue with strings in TextToSpeech action
A change in EventGhost starting with 0.5.0-beta1 seems to have changed how the
string class is handled, breaking the Speech plugin.

Previous to this change attempting to add the TextToSpeech action fails
with:
"AttributeError: class TextToSpeech has no attribute 'buttonInsertTime'"

Previous to this change the TextToSpeech action did not display the correct
name and had no description.