Adds Python formatter
Adds PrettyPythonPrint to eg.Utils.

This is use to format an attribyte name, attribute value to be written to a file. It formats name = value with indent to conform to 80 character line length it will split tuples and lists but also will wrap long strings with brackets and break the string into multiple lines