Adds variable parameter to eg.Environment.Get
Adds parameter variable to the Get method in eg.Environment. it completes the convience method. It is defaulted to None. so if no variable is specified it will do what it was originally programmed to do and that is return a sorted list of the variables. but if a variable name is specified it will return that variable if it has been set otherwise it will return None.