Fixes System.MonitorPowerOn (Windows 8 & 10)
Fixes not being able to turn the monitor on from the System.MonitorPowerOn action when running Windows 8 or Windows 10.

Changed MonitorPowerOff and MonitorStandby to use the same mechanism.

SendMessage from WinApi/Dynamic is actually an alias for ctypes.WinDLL('user32').SendMessageW. and as this does not cause an issue with windows 7 it seems to not work for Windows 8 or 10 for turning the monitor back on

so the change I made uses pywin32 instead of ctypes and it powers the monitor on properly