Removes automatically adding GUID
Changed the automatic entry of the guid for a simple dialog providing one. This is due to the complexities and issues that can come up fith variables being passed into the call to eg.RegisterPlugin. There are plugins that I have seen where a variable version is passed. I am sure there are others.

Moved all dialog messages and titles to eg.Text so they will be available to be translated