Adds checking to see if dialog has any unsaved data
Adds checking to see if the open dialog is a standard EG dialog. and if so checks to see if the dialog has any undaved data and if it does raise the dialog and do the blink and beep thing. and if it does not close the dialog.

If the dialog is not a standard EG dialog.  to raise the dialog and do the blink and beep thing

Removes the message dialog informing the user of open dialogs as this has become obsolete with the addition of unsaved data checking.

Removes caption and message for the message dialog from Text.py