Fixes not being able to close EG if a traceback occurs in a dialog
If a traceback occurs during the creation of the dialog the FinishSetup never gets to run there for is unable to set the flag setupFinished. even tho a notification has already been sent out and the dialog has been added to the openDialogs list in the mainFrame. So what I have done is removed the checking of setupFinished in the TaskletDialog.ProcessingTask this will allow for the dialog to be destroyed and the notification to be sent to remove the dialog. wether or not the the dialog finished setup.