Fixes thread timing issue if a dialog has a traceback
If a dialog has a traceback  that occurs during the creation of the dialog. there is a timing issue in the threads that causes extra tracebacks to occur.

this is the setup..

tree Item gets selected. causing a eg.Notify routine to run. the dialog gets created. traceback occurs. then the dialog gets destroyed. sending another Notification to remove the dialog from this list. and the treeItem gets deleted in a seperate thread somewhere along these lines it causes the item.parent in tree Item to become None because it was deleted.. because there are 2 threads running one that deleted the item and the other that is still working on selecting the item the one that is selecting the item duns across NoneType errors. 
so these 2 changes account for that possibility and will remove the additional tracebacks that really do not need to be shown when an item gets deleted due to a traceback from a dialog.