Adds ability to change polling time
Adds a plugin config option to set the polling interval by setting a lower number this will remove some cpu load time.

Changed the use of Time.sleep to threading.Event to allow the thread to exit in a clean manner

Added an action to dynamically change thepolling interval with a timer to reset it back. This would be used if you were looking for a chain of things to happen and didn't need to have it running at full speed all the time. and the Timer to set it back is a convience thing. if the Timer value is set to 0 it will remain at the new value.