ReAdds work around for traceback if min value > 0
ReAdds the setting of the vaue other parameters until after the control has been initilized.

We had thought this issue fixed in the new version of wx and also tested and it worked properly. but there is a random issue that can show up. and this issue has not been fixed by the wx team. So the work around had to be put back in