Removes minimize preview, combines code for setting pane properties
Removes minimize preview due to a bug which causes the preview to open over the restore button where you have to hover the mouse to view the preview. and since the preview is now under the mouse the preview closes. and once it closes the button is then again under the mouse so it open back up again.. and so on and so forth

Combines code for setting of the pane properties.