Changes the WinUsbWrapper.dll file
Changes the WinUsbWrapper.dll file to the one from EG 0.4.
This is done because the source files for the dll are identical between versions but the file sizes are not the same.

There also appears to be a descrepancy between modified dates. But i am unsure of why I have different modified dates for the EG 0.5 version of it.