Currects event.payload
This modification will properly handle a None payload. where as EG at the present time does not display any payload data in the log if it is None. There are instances where a None payload is an actual payload value. and should be displayed. 

This is probably going to break peoples code if they specifically look for None as a non value in the payload. The reason why is event.payload is only created if a payload value has been set. I have done this through the use of a class so unless the payload value somehow gets set to that class othern then by default this will give true payload log printing as well as payload data if it exists.