Adds GetDevices action
Adds a new action GetDevices. this is a very talented class, it supports the use of wildcards by use of fnmatch.fnmatch '?', '*'. And since searching through all of the devices on a computer can take a wee bit of time. and of the expensive fnmatch we want to make this as quick as possible. If you use the dialog it will return a dict with the device type and the pattern to search for. this way I only have to queue up the system for that one specific device type. this same method can be used from script as well. but I have also set up the script so you can just use the device type as the actual keyword. less any spaces of course. the device types are the containers in the dialog for the different devices.