Fixes missed device/volume notifications
Volume events are being generated now
Removed return from device lookup. so now all drivers for a specific device will generate an event. as an example my webcam has a microphone in it. so instead of just getting an event for the webcam i also get one for the microphone as well.