Fixes help and other bugs, reformats code
Fixes the data corruption with help. this issue was with wx.iewin_old and the activeX control it uses. I converted the help files to markdown and used eg.HtmlWindow

Sets the size for the config dialog
Positions the sash
Adds startup help file to the root item in the tree
Fixes expanding and collapsing of the tree items.
Code cleanup.
Added docstrings