Changes eg.Cli to use argparse
Changes eg.CLI from using sys for parsing command line arguements to using argparse.
I have also added error reporting in the event of an incorrect argument. This is done through the use of a wxDialog because py2exe overrides stdout.

There is one big change and that is the -h -hidden is not longer supported. this is because of -h -help, so the -h -hidden has been changed to -mg -minimizegui since this is really what it does in the first place. I do plan on hadding a -ng -nogui option to completely hide EG except for having it in the process list. This is not a large issue due to someone using the -h is going to bring up the help dialog.. and they can read it and discover the change

I have also added -- switches as well.

Todo: Reformat the help