Fixes broken clipboard chain in Windows 10
Windows 10 has become somewhat fickle about having the same thread that registered for a notification also be the thread that unregisters.  so I have moved the notification of a clipboard chain change to the MainMessage receiver thread. but do to it's proximity to closing the actual window that receives the notifications I have put a threading lock object into place to halt the closing of the window until the chain has been repaired,