Remove wiki content from documentation
The documentation pages "Controlling your living room with EventGhost" and "X10 Remotes" are duplicates of pages in the EventGhost wiki (http://www.eventghost.net/mediawiki/index.php?title=Main_Page). It's a very bad idea to have duplicate content and these pages are more more appropriate for the wiki. I have merged all updates to the content of these pages to the wiki already.

The "X10 Remotes" was not part of the documentation index so its removal will have minimal effect on the documentation.