Changes restarting EG to using a Named Pipe
Removes the use of restarting EG to start an elevated eventghost. Instead If an EG instance is already running it will send the command via a named pipe to the currently running EG session. 

I also added the ability to hide EG if it is already running using the -h or -hide cli switch