Exposes EG API to Named Pipe
I exposed the whole eg API to the Named Pipe. the name of the pipe is \\.\pipe\eventghost_pipe. the structure of the data that has to be sent through the pipe is as follows

"eg.PluginInstall.Import, (PLUGIN_FILE,)"

or

"eg.PluginInstall.Import, [PLUGIN_FILE]"

or 

"eg.PluginInstall.Import, dict(filepath=PLUGIN_FILE)"

or 

"eg.PluginInstall.Import, {'filepath': PLUGIN_FILE)"

the information sent through the pipe has to be a string. 
so the module in EG you want to access along with a tuple or a list of the positional arguments or a dict of the keywords and data to be passed to that keyword.
All functions/classes/methods that are ran get run in the main thread and arguments get sent using * if a tuple or a list so make sure they are in the proper order. and ** if it's a dict so if you need to pass a dict to a positional arguments make sure you do this properly and surround it with ()'s

There has been some checking done with the implimentation of this. if you do the following

eg.PluginInstall.Import('SOME_PLUGIN')

it will run the command as 

eg.PluginInstall.Import()

so make sure you separate the arguments from the function/method/class

if you specify no arguments it will pass no arguments.

The named pipe will run any function/method/class you send to it. it will not set variables so if you need this done best to make a function to do it and set that function into eg.globals then you will be able to set the variables you want.

I have done my best to protect this from running any kinf of code that is not supposed to be there. If you find a hole please let me know.