Adds variable to eg.Config for turning on split logging
if split logging is to be used the config.py file needs to be edited and the line splitDebuggingLog = True needs to be added to it in order to turn on this feature.

added for printing the header information to the original Log output. 
removed the computer name from the header.