Fixes 2 instances of EG running
After using the dropdown menu to restart EG and then using the -restart cli switch there would be 2 running instances of EG. I made it so that appmutex is not used to check for a running instance of EG. instead i used the named pipe. I added a ping, pong test to check for a currently running instance. by not using appmutex to check i am able to use sys.exit to exit the instance  if there is already a running instance. and this seems to have fixed the issue.