Fixes double instance when debugging is turned on
EG would start a second instance if debugging was turned on. when the --restart cli switch was used.

It turns out that because of my excessive exception catching it would cause everything to get bypassed like nothing was wrong when in fact there was a pipe error that was being triggered. and this was error 231 pipe busy. and this would happen if messages were sent back to back. So not it does a check for this and if the pipe is busy it will continue to try and send the message until it can,