Changes XmlIdLink to eg.GUID (#264)
To turn it on you would open the add plugin dialog and left click once in the upper right hand corner of the dialog (client area) and then click once in the lower left hand corner of the dialog (client area) there is a 20x20 pixel target for each of the spots so the cance of someone doing this accidentally is probably not going to happen. But in the event it does a message box asking if you want to enable or disable it pops up.