Removes ability for the user to have an un-configured action
Disabled the OK button in Cinfigure dialogs if the user has not made any changes. This makes it so that when a new action is added it will not create the tree item unless the action has been configured. It will only enable the ok button once the apply button has been pressed