Adds convience module to get specific monitor attributes
Adds DesktopDisplay.py

methods:
Refresh() - refreshes the display list. this automatically gets called when EnumDisplays() or GetDisplay() is called.

EnumDisplays() - returns a list of displays

GetDisplay() - you can either pass this the monitor number or monitor name and this will retrun a DesktopDisplay.Display instance

attributes
displays - this is a list of all of the available displays that are attached
size - this is the virtual size of the desktop display
displayCount - number of connected displays
isTouchScreen - take a guess
isMultiTouch - take a guess
maxTouchCount - maximum number of simultanious touches if the display is touch capable otherwise 0

DesktopDisplay.Display: 
this class will be returned as a reprsentation of a specific display that is attached.

attributes:
blackout - set True/False to blackout the display
x - upper left corner of the display pixel position relative to the width of the desktop
y - upper left corner of the display pixel position relative to the height of the desktop
w - display pixel resolution width
h - display pixel resolution height
primary - True/False if the display is the primary display
number - display number, this the the number seen in the windows screen resolution dalog
name - nameof the display. this is the name seen in the screen resolution dialog
size - this is the size of the display returned as (w, h)
pos - this is the upper left corner pixel position of the display returned as (x, y)

If for example you have 3 displays attached and you disconnect display #2 this will automatically cause Windows to reorder the displays this will also reorder as required. 

If the screen is blacked out during this reordering it will remove the blackout.

If you change the display position when the screen is blacked out, the blackout should remain