And the fight continues
More code compaction
Created the file with norepad first then copied code into the file. I believe this will fix the BOM problem with code climate. Event thos I have my IDE set to write the BOM it won't because of the magic encoding comment, this turns off the BOM and CodeClimate either doesn't handle or isn't configured to use the encoding comment line. But if i make the file outside of my IDE and then edit it i believe this will keep both the BOM while still having the encoding comment line. tho both are not needed because if there is no encoding line then BOM is used.