Adds Event EventGhost.Started
Adds  EventGhost.Started event to take place before the running of the Autostart macro. This is done so that if the user needs to set into place any globals that a script may depend on it can be done before the possibility of a plugin starting and generating an event before the Main.OnInit gets triggered.

There is a payload with this event that is an instance of a Time(datetime.datetime) object I subclassed the datetime.datetime class to override the __str__ and __repr__ methods so it would produce a nice date in the log. other then that it functions exactly the same as a datetime.datetime object