Fixes tracebacks
Correction: event name is EventGhost.Startup
Removed subclassing of the datetime.datetime object as this caused some issues. The object returned is a class that stores the datetime.datetime instance and will returns any attributes to has correctly other then __repr__ and __str__

I also removed adding the execution of the event into the action thread. The Autostart.Execute() was already being run in the action thread. and this caused a hang because it would wait for the event to get run when the Execute was already being run.