Adds event EventGhost.LoggedInUser.USERNAME
Adds the event EventGhost.LoggedInUser.USERNAME on startup
This event is the first event hat will occur. This will allow for macros to be run depending on the username but also if they are administrator or not.

There is a payload that is a dict(IsLocalAdmin=bool, IsDomainLogin=bool, IsDomainAdmin=bool)
IsDomainAdmin will only be added if IsDomainLogin is true
example:    EventGhost.LoggedInUser.Administrator {'IsLocalAdmin': True, 'IsDomainLogin': False}