Reworks how the task is made
This adds a task for when windows boots. it will login and run eventghost. User credentials are required for this to function and you are prompted to enter your windows password for the user you are logged in with.

It also adds a task to run eventghost for specific usernames.

How all of this works is If you click on start EG with user User1 then the User1 account is used to run eventghost when the system boots. if you then login  with User2 and set EG to start when User2 logs in it will remove User1 from the boot task and place User2 in it's place and then it will add another task to start EG when User2 logs in. So you would end up with 3 tasks. one for boot one for User1 login and one for User2 login.

If the Windows version is > XP EG will use this new method of starting up instead of making an icon in the startup folder. If you change the startup option at any point and there is a startup shortcut for EG it will be removed.