Fixes eg.EqualizeWidths
eg.EqualizeWidths was coded to accept a single iterable of controls. I have found this very limiting.

So what I have done is it uses *args and you can pass single controls as positional areuments, or you can pass nested lists/tuples or a combination

eg.EqualizeWidths(ctrl1, ctrl2, ctrl3, ctrl4) works the same as eg.EqualizeWidths((ctrl1, ctrl2, ctrl3, ctrl4))


eg.EqualizeWidths((ctrl1, ctrl2, ctrl3, ctrl4), (ctrl5, ctrl6, ctrl7, ctrl8)) this will set the widths the same for all of the controls in tuple 1. and then it will set the widths the same for tuple 2. tuple 1 will not have the same widths as tuple 2

eg.EqualizeWidths(ctrl1,(ctrl2, ctrl3, ctrl4), ctrl5, (ctrl6, ctrl7, ctrl8)) the tuples works the same as above but in this example you see ctrl1 and ctrl 5 are not in a tuple. those 2 controls will have matching widths but they will not match the widths of tuple 1 or tuple 2