Equalizes the widths of the StaticText controls
When using the AddLine method of eg.ConfigPanel this causes the labels and controls to be rendered in a not so pleasing manner. I have added the equalizing of even numbered columns if the control is a wx.StaticText instance and if there are more column items in that specific row. it will add these instances into matching groupes based on column number. and each grouping of controls then get passed to EqualizeWidths.

This produces a much nicer looking dialog.