Adds better handling of default parameters
I added better handeling of the default parameters. removed somethings that were redundant. and also added exceptions for the value being lower then the min or higher then the max.

I have still not been able to locate the bug that is in the wx.masked control I read through the code and I do not see where one could be. I would need a Traceback of the bug to be able to pin it down. I am unable to locate a forum posting for the bug so I really have nothing to go by.