Updates builder
I have added having a command line switch as well as a check box in the
GUI to have the builder download any necessary deps. Everything except
for stackless. I have also added support for egg files. specifically
checking to see if the egg files are zips. In the event they are an
error will be thrown. If you launch the builder ith the -install switch
it will also download and install wx (if necessary) before it launches
the GUI when the GUI is used it automatically sets the verbose logging.
This is due to the text ctrl i added to view the log output from the GUI

I also added a fix for jinja 2. so now we have the most recent version
of sphinx added the the deps. someone has decided to pick up development
of ctypeslib. it is called ctypeslib2 but carries the same module name
as before, So this has also been added to the deps. All other packages
have also had their versions updated except for stackless python.