Rewrte of named pipe
Removes use of pywin32 for handling the named pipe.

Only uses python std lib

Changes handling of commands so only commands directed to eg.document or eg.mainFrame will be run in the main thread.

Changes the cosing of the Pipe so the client has to initiate the close by sending CLOSE. this allows the pipe connection ro remain open for further commands to be sent over the pipe without the need to create a new connection.

Changes the pipe so a pipe name can be supplied. This is so that the named pipe code is not dedicated to only the use of EG. it can now be used in plugins or a script if needed.

Changes the pipe instance checker. I have created a seperate function is_pipe_running that allows the user to supply a pipe name to check for the instance of a named pipe.

Changes the exception creation in the command processor to no longer throw exceprions. Instead if an error occurs the returned value will be one of the following.

CommandFormatError
ParameterFormatError
CommandMalformedError
CommandNotFoundError

There is alow the following 2 errors that can be returned. These are only returned in the event of the command that was executed not returning within 5 seconds. because we run the command in 2 different threads the main thread or some other thread the returned error values will reflect that.
MainThreadHung
UnableToProcessCommand