Removes exception classes
Removes the different exception classes in favor of using a single class. 

Adds a function to print to the error log instead of throwing an exception in the send_message function.  When the write_error function gets called it will disable the message box that is displayed by EG if something is written to the error log on EG startup. 
The reason these  things are done is if there is an error during transmission and this function is being used to pass cli commands to a running instance of EG this will cause a second eventghost process to get stuck. in turn causing the first to hang.