Adds read method to pipe instance
This adds a read method to a connection instance for the pipe. This read method only becomes usable if the pipe name is not "eventghost", This makes the pipe server class fully usable outside of the core. The read method contains all locks and wait objects to make it thread safe. the read method will wait forever if there is no data availale so be sure to make a thread to handle reading and also make the thread a daemon thread as there will be no way to end the thread when EG shuts down. so make sure that the thread has no oerations that does that would require a safe close of an object.