I finally did it!!!!
OK so i beat the snot out of this thing once again and i believe this time i might have the issues sorted out.

I ran 250 consecutive CLI events and it didn't even bat an eyelash at it. finished all 250 in just under 4 seconds. that is with EG doing all of the subprocess leg work as well. so not to shabby.

then i decided to give it a real storm trooping and went with 250 concurrent events. and it had similiar performance data. and it completed all 250 events


then i di a 400 event flip flop so 100 consecutive, 100 concurrent, 100 consecutive and then 100 concurrent
this took 6 seconds to complete avg of 15ms each. 

there was never an orphaned EG process. it worked smooth as glass