Improves performance when using  WinApi.SendKeys.SendRawCodes1
The old mechanics of this method would send one key at a time and then wait for the notification to get processed by the receiving window. The user32.SendInput function allows you to send multiple keystrokes at a time. 

I have modified the code so it now sends all of the keystrokes in a single notification and then waits only once. The difference in performance is astonishing.

The xml below is a way of testing this. It is a very strange way you have to go about getting the SendKeys action to use this method for sending keystrokes. You need to use Find Window to locate the window. then you have to bring the window to the front (this gives it focus) then in a python script you need to delete the contents of eg.lastFoundWindow, then run the Send Keys action. This will send the keys using this method. 

I suggest using the code below on an unaltered version of EG so you can see the speed. the code below requires you to have notepad open as this is what it is going to send the keys to. The run the same test using this modified version.

    <?xml version="1.0" encoding="UTF-8" ?>
    <EventGhost Version="WIP-2018.05.25-17.43.19">
        <Macro Name="Find Window: notepad.exe" XML_Guid="{EF46CE14-DFF8-499C-9B35-840D5C348282}" Expanded="True">
            <Action XML_Guid="{15930DB5-AEB3-4F57-A0B0-840161990957}">
                Window.FindWindow(u'notepad.exe', None, None, None, u'Edit', 1, False, 0.0, 0)
            </Action>
            <Action XML_Guid="{0C9CF2AF-71AA-4992-8877-41DA592545BD}">
                Window.BringToFront()
            </Action>
            <Action XML_Guid="{CD0A4449-AA83-482B-A9D5-661DF8B0658F}">
                EventGhost.PythonScript(u'del eg.lastFoundWindows[:]')
            </Action>
            <Action XML_Guid="{63EC7FC7-320E-4602-A93F-BA3B9A6A032B}">
                Window.SendKeys(u'This is a test.This is a test.This is a test.This is a test.This is a test.This is a test.This is a test.This is a test.This is a test.This is a test.This is a test.This is a test.This is a test.This is a test.This is a test.This is a test', False, 2)
            </Action>
        </Macro>
    </EventGhost>