Adds using meth SendRawCodes1 without deleting eg.lastFoundWindows
I added a test of the window handle being passed to WinApi.SendKeys.__call__ against the window handle that has focus. if the 2 handles match or if there is no window handle passed SendRawCodes1 will get used.
(in most cases a user is going to want to send keys to the window they are looking at so this will be a nice addition especially if the text that is being input is lengthy. 

with how the code was before because of the use of the PostMessage function if a string of text is being sent to the same window that is currently in focus and the user types something you will end up with the text not being properly inputted.

The SendInput function inserts the events in the INPUT structures serially into the keyboard or mouse input stream. These events are not interspersed with other keyboard or mouse input events inserted either by the user (with the keyboard or mouse) or by calls to keybd_event, mouse_event, or other calls to SendInput. This is a much better mechanism to be able to send the keystrokes. 1 because of it's speed and 2 because it cannot be interfered with.