Improves speed of SendRawCodes2
This modification greatly improves the speed of sending keys to a background window. 
  
The calls to WaitForInputProcessed were done for each and every single key press. This is not necessary. There are 2 times when a wait needs to be performed

if the current key code block contains more then a single key code. or if the block that is next in line has more then a single key code. This test has to be performed  after a key press and after a key release. not for each key in a key code block. and not if there is only a single key.


there are 2 mechanisms at work. the sending of a key code and the setting of the keyboard state. when the shift key gets set in the keyboard state change it has a global affect so if a key code that has been sent has not been processed and a change to the keyboard state takes place then the  outcome of the key code once it does get processed will be different. This is the purpose for waiting until it does get processed before continuing.