Fix for jinja2 > 2.8
This is a fix for jinja2 > 2.8 SyntaxError when p2exe is byte compiling the files. The traceback occurs because newer versions of jinja 2 have files that include asyncio code.

The way i wrote this fix is it will check the source for any reference to an import of asyncio and if founf it will check the python version. if asyncio is not supported for that version it will return an empty byte compiled file