Adds EventGhostEnduringEvent class
Currently there is no means of knowing what kind of event has been triggered. This modification fixes that problem. A different icon is now used to identify an enduring event from a standard one. This will make it far easier for the user to implement looping code to perform actions in succession.  

I was pretty crafty in how i display the icon. because i not only wanted to have the icon changed in the log but also in the tree as well. once an event has been triggered and set as an enduring event it will change the icon in the tree. when an event is added to the tree it will check and see if that event has been flagged as an enduring event. if so it will change the new events icon accordingly. an XML attribute of EventType is added to the save data to identify the icon to use.